feat(dashboard): slot row drawer integration with patient name fallback
- Extract ScheduleSlotRow as dedicated component with staggered animation,
  left-accent border, status chip, live progress bar, and click-to-open drawer
- Lift AvailabilityWeekDrawer to Dashboard page level (outside widget)
- Map full IWeekSlot (id, _id, availabilityDayId) in useDashboardSlots so
  the drawer receives all fields it needs
- Fall back to slot.patientName in AvailabilityWeekDrawer while
  appointmentDetails loads, so avatar and title render immediately
- Fix Avatar: src="" was suppressing initials; use src={undefined} when
  no photo; guard empty names with optional chaining to avoid "undefinedundefined"
- Add min-width:0 to BentoTileRoot to prevent grid overflow on mobile
- Remove slot-row styles from ScheduleWidget.styles (moved to ScheduleSlotRow.styles)
