+++ src/components/dashboard/widgets/schedule-widget/ScheduleSlotRow.types.ts
@@ -1,0 +1,8 @@
+import type { IWeekSlot } from '@psycron/pages/availability/week/AvailabilityWeekPage.types';
+
+export interface ScheduleSlotRowProps {
+	index: number;
+	onClick: (slot: IWeekSlot) => void;
+	showDate?: boolean;
+	slot: IWeekSlot;
+}
